[Refactor] Move HttpRequest and HttpTask to standalone module
