+++ app/build.gradle
@@ -194,6 +194,8 @@ dependencies {
 
     implementation project(':firebase')
     implementation project(':feature-tabs')
+    implementation project(':HttpRequest')
+    implementation project(':httptask')
 
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk8:${Versions.kotlin}"
